rh: Treat restricted CQ close as undeliverable
When an application terminates, restricted get packets will timeout.
This is resulting in the CQ close logic being triggered. The issue is if
the restricted get is large (e.g, 1 GiB), the RH will cancel packets at
a rate of 64 per second. This will result in the get taking 524,288
seconds to be cancelled.

Treating the CQ close condition as undeliverable results in the parked
NID logic being applied. This adjusts SPT timeout and packets inflight
resulting in the restricted get being cancelled much faster.

NETCASSINI-6929

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>