RH: Honor user provided retry_intervals
Commit a73bf6d which deprecated timeout_backoff_multiplier and
timeout_backoff_factor in favor of retry_intervals inadvertently removed
a line to store and use user provided retry_intervals. Reintroduce that
code.

Updated related logging and comments

1) Print out each retry_interval that will be used - this may be the
defaults or what the user provided.

2) We read retry_intervals in from the user using seconds as the unit.
The values are stored as usec values for use by RH timers. If we see an
error, print the interval in the same unit provided by the user.

3) Add a warning in cxi_rh.conf that a change to the TCT Epoch will require
modifying retry_intervals as well.

Signed-off-by: Amith Abraham <amith.abraham@hpe.com>